Abstract runner.
