+++ bin/builder.js
@@ -12,6 +12,7 @@ env.updateNodePath("node_modules/builder-react-component/node_modules");
 // Infer task to run
 var Task = require("../lib/task");
 var task = new Task();
+log.info("task", task.toString());
 
 // Detect help
 if (task.action === "help") {
@@ -23,23 +24,7 @@ if (task.action === "help") {
   process.exit(0);
 }
 
-var cmd = task.getCommand();
-
 // Run the task
-// TODO: Decompose to `lib/runner.js`
-var exec = require("child_process").exec;
-var proc = exec(cmd, {
-  env: env.env
-}, function (err) {
-  if (err) {
-    log.error("error", "Failure: " + cmd);
-    process.exit(err.code || 1);
-  }
-});
-
-proc.stdout.pipe(process.stdout, { end: false });
-proc.stderr.pipe(process.stderr, { end: false });
-
-// TODO: Move around logger.
-log.info("task", task.toString());
-log.info("exec", cmd);
+var runner = require("../lib/runner");
+var cmd = task.getCommand();
+runner.run(env, cmd);
